feat: 마이페이지 아키텍처 리팩토링 및 신규 디자인 토큰 적용 close #VO-882
