+++ frontend/src/app/community/components/CommunityCard.module.css
@@ -8,10 +8,11 @@
   /* 24px */
   gap: var(--space-16);
   /* 16px */
-  width: 410px;
+  width: 100%;
+  max-width: 100%;
   height: 495px;
-  background-color: var(--color-surface);
-  /* #F3F4F6 */
+  background-color: var(--color-text-white);
+  box-shadow: 0 0 12px rgba(0, 0, 0, 0.08);
   border-radius: var(--radius-xl);
   /* 24px */
   box-sizing: border-box;
@@ -21,7 +22,7 @@
 
 .card:hover {
   transform: translateY(-2px);
-  box-shadow: var(--shadow-md);
+  box-shadow: 0 0 20px rgba(0, 0, 0, 0.12);
 }
 
 .header {
