+++ frontend/src/app/community/components/CommunityListCard.module.css
@@ -6,7 +6,8 @@
   gap: var(--space-16);
   width: 100%;
   max-width: 100%;
-  background-color: var(--color-surface);
+  background-color: var(--color-text-white);
+  box-shadow: 0 0 12px rgba(0, 0, 0, 0.08);
   border-radius: var(--radius-xl);
   box-sizing: border-box;
   text-decoration: none;
@@ -15,7 +16,7 @@
 
 .card:hover {
   transform: translateY(-2px);
-  box-shadow: var(--shadow-md);
+  box-shadow: 0 0 20px rgba(0, 0, 0, 0.12);
 }
 
 .header {
