+++ frontend/src/app/mypage/communities/page.tsx
@@ -1,4 +1,4 @@
-import styles from '../page.module.css';
+import styles from './page.module.css';
 
 export default function MyCommunitiesPage() {
   return (
