+++ frontend/src/app/mypage/events/page.tsx
@@ -1,26 +1,12 @@
 'use client';
 
-import React, { useState, useEffect, useCallback } from 'react';
+import React from 'react';
 import { useRouter } from 'next/navigation';
 import Sidebar from '@/components/layout/Sidebar';
 import { Card, CardGrid, Tabs, Pagination, InputField } from '@/components/ui';
 import { ReviewModal } from '@/components/modal';
 import styles from './page.module.css';
-
-interface LectureItem {
-  orderId: number;
-  status: string;
-  title: string;
-  organizer: string;
-  dateTime: string;
-  location: string;
-  price: number;
-  eventId: number;
-  thumbnailUrl?: string;
-  categoryId?: number;
-}
-
-const ITEMS_PER_PAGE = 8; // 2열 × 4줄
+import { useEvents } from './useEvents';
 
 const TAB_OPTIONS = [
   { value: 'upcoming', label: '진행 전' },
@@ -36,115 +22,22 @@ const CATEGORY_MAP: Record<number, string> = {
 
 export default function MyPage() {
   const router = useRouter();
-  const [activeTab, setActiveTab] = useState('upcoming');
-  const [currentPage, setCurrentPage] = useState(1);
-
-  const [lectures, setLectures] = useState<LectureItem[]>([]);
-  const [wishlistSet, setWishlistSet] = useState<Set<number>>(new Set());
-  const [totalPages, setTotalPages] = useState(1);
-  const [loading, setLoading] = useState(false);
-
-  // 리뷰 모달 상태
-  const [reviewModalOpen, setReviewModalOpen] = useState(false);
-  const [reviewTarget, setReviewTarget] = useState<{ eventId: number; title: string } | null>(null);
-
-  // 이미 리뷰를 작성한 이벤트 ID 추적
-  const [reviewedSet, setReviewedSet] = useState<Set<number>>(new Set());
-
-  // 백엔드 API 호출
-  const fetchOrders = useCallback(async (tab: string, page: number) => {
-    setLoading(true);
-    try {
-      const res = await fetch(
-        `/api/orders/me?tab=${tab}&page=${page - 1}&size=${ITEMS_PER_PAGE}`
-      );
-      if (res.ok) {
-        const json = await res.json();
-        const data = json.data;
-        const validLectures = (data.content || []).filter(
-          (item: any) => item.status === 'PAID' || item.status === 'REGISTERED' || item.status === 'PENDING'
-        );
-        const mappedLectures: LectureItem[] = validLectures.map((item: any) => ({
-          orderId: item.orderId,
-          eventId: item.eventId,
-          title: item.eventTitle,
-          status: item.eventStatus?.label || item.status,
-          organizer: item.organizer || "알 수 없는 호스트",
-          dateTime: item.eventStartDate ? new Date(item.eventStartDate).toLocaleString('ko-KR', {
-            year: 'numeric', month: '2-digit', day: '2-digit', hour: '2-digit', minute: '2-digit'
-          }) : "-",
-          location: item.location || "-",
-          price: item.amount,
-          thumbnailUrl: item.thumbnailUrl || undefined,
-          categoryId: item.categoryId || undefined,
-        }));
-        setLectures(mappedLectures);
-        setTotalPages(data.totalPages || 1);
-
-        // 종료 탭일 때 각 이벤트의 리뷰 작성 여부 확인
-        if (tab === 'completed') {
-          const eventIds = [...new Set(mappedLectures.map(l => l.eventId))];
-          const reviewed = new Set<number>();
-          await Promise.all(
-            eventIds.map(async (eid) => {
-              try {
-                const rRes = await fetch(`/api/events/${eid}/reviews/can-review`);
-                const rData = await rRes.json();
-                // can-review가 false면 이미 리뷰를 작성한 것
-                if (rData.success && rData.data === false) {
-                  reviewed.add(eid);
-                }
-              } catch { /* ignore */ }
-            })
-          );
-          setReviewedSet(reviewed);
-        }
-
-        // Fetch wishlist IDs
-        try {
-          const wlRes = await fetch('/api/wishlists/me?size=100');
-          if (wlRes.ok) {
-            const wlData = await wlRes.json();
-            if (wlData.data && wlData.data.content) {
-              const ids = new Set<number>(wlData.data.content.map((item: any) => item.id));
-              setWishlistSet(ids);
-            }
-          }
-        } catch (e) {
-          // ignore
-        }
-      } else {
-        setLectures([]);
-        setTotalPages(1);
-      }
-    } catch (err) {
-      console.error('Failed to fetch orders:', err);
-      setLectures([]);
-      setTotalPages(1);
-    } finally {
-      setLoading(false);
-    }
-  }, []);
-
-  // 탭 또는 페이지 변경 시 API 재호출
-  useEffect(() => {
-    fetchOrders(activeTab, currentPage);
-  }, [activeTab, currentPage, fetchOrders]);
-
-  const handlePageChange = (page: number) => {
-    setCurrentPage(page);
-    window.scrollTo({ top: 0, behavior: 'smooth' });
-  };
-
-  const handleTabChange = (value: string) => {
-    setActiveTab(value);
-    setCurrentPage(1);
-  };
-
-  // 리뷰 작성 가능 여부 판단
-  const canWriteReview = (eventId: number) => {
-    return activeTab === 'completed' && !reviewedSet.has(eventId);
-  };
+  const {
+    activeTab,
+    currentPage,
+    totalPages,
+    lectures,
+    loading,
+    wishlistSet,
+    reviewModalOpen,
+    reviewTarget,
+    handleTabChange,
+    handlePageChange,
+    canWriteReview,
+    openReviewModal,
+    closeReviewModal,
+    handleReviewSuccess,
+  } = useEvents();
 
   return (
     <div className="container-sidebar">
@@ -193,8 +86,7 @@ export default function MyPage() {
                   }
                   onActionClick={() => {
                     if (canWriteReview(lecture.eventId)) {
-                      setReviewTarget({ eventId: lecture.eventId, title: lecture.title });
-                      setReviewModalOpen(true);
+                      openReviewModal(lecture.eventId, lecture.title);
                     } else {
                       router.push(`/events/${lecture.eventId}`);
                     }
@@ -202,8 +94,7 @@ export default function MyPage() {
                   secondaryActionText={activeTab === 'enrolled' ? '리뷰 작성하기' : undefined}
                   onSecondaryActionClick={() => {
                     if (activeTab === 'enrolled') {
-                      setReviewTarget({ eventId: lecture.eventId, title: lecture.title });
-                      setReviewModalOpen(true);
+                      openReviewModal(lecture.eventId, lecture.title);
                     }
                   }}
                 />
@@ -211,7 +102,7 @@ export default function MyPage() {
             </CardGrid>
 
             {!loading && lectures.length === 0 && (
-              <p style={{ color: 'var(--color-text-gray-500)', textAlign: 'center', width: '100%', padding: 'var(--space-48) 0' }}>
+              <p className={styles.emptyState}>
                 해당 탭에 이벤트가 없습니다.
               </p>
             )}
@@ -229,15 +120,10 @@ export default function MyPage() {
       {reviewTarget && (
         <ReviewModal
           isOpen={reviewModalOpen}
-          onClose={() => setReviewModalOpen(false)}
+          onClose={closeReviewModal}
           eventId={reviewTarget.eventId}
           eventTitle={reviewTarget.title}
-          onSubmitSuccess={() => {
-            setReviewModalOpen(false);
-            // 리뷰 작성 완료 → 해당 이벤트를 reviewedSet에 추가하여 버튼 숨김
-            setReviewedSet(prev => new Set(prev).add(reviewTarget.eventId));
-            setReviewTarget(null);
-          }}
+          onSubmitSuccess={handleReviewSuccess}
         />
       )}
     </div>
