+++ frontend/src/app/mypage/profile/page.module.css
@@ -130,3 +130,31 @@
   flex-flow: row wrap;
   gap: var(--space-12);
 }
+
+.categoryPill {
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  padding: var(--space-8) var(--space-16);
+  border-radius: var(--radius-full);
+  border: 1px solid var(--color-text-gray-300);
+  background: transparent;
+  color: var(--color-text-gray-700);
+  font: var(--text-body-medium);
+  cursor: pointer;
+  transition: all var(--transition);
+}
+
+.categoryPill:hover {
+  background: var(--color-surface);
+}
+
+.categoryPillActive {
+  background: var(--color-primary);
+  color: var(--color-text-white);
+  border-color: var(--color-primary);
+}
+
+.categoryPillActive:hover {
+  background: var(--color-primary-hover);
+}
