+++ frontend/src/app/mypage/wishlist/page.module.css
@@ -1,4 +1,4 @@
-/* mypage/page.module.css */
+/* src/app/mypage/wishlist/page.module.css */
 
 .content {
   display: flex;
