+++ frontend/src/components/layout/Footer.module.css
@@ -8,7 +8,7 @@
   gap: var(--space-100);
   width: 100%;
   height: 310px;
-  background-color: var(--color-text-gray-50); /* #F9FAFB */
+  background-color: var(--color-text-gray-50); 
 }
 
 /* Frame 70 (링크 영역 레이아웃) */
