+++ frontend/src/components/layout/Sidebar.module.css
@@ -6,8 +6,7 @@
   padding: var(--space-24) var(--space-12);
   gap: var(--space-8);
   width: 240px;
-  background-color: var(--color-text-gray-50);
-  border-radius: var(--radius-xl);
+  background-color: transparent;
   flex: none;
   flex-shrink: 0;
   box-sizing: border-box;
@@ -42,11 +41,12 @@
 }
 
 .selected {
-  background-color: var(--color-text-gray-900);
+  background-color: var(--color-primary);
   color: var(--color-text-white);
 }
 
 .selected:hover {
+  background-color: var(--color-primary-hover);
   color: var(--color-text-white);
 }
