+++ frontend/src/styles/globals.css
@@ -116,7 +116,7 @@ select {
 .sidebar-content {
   flex: 1;
   /* 오른쪽 본문 영역만 확장 */
-  max-width: 848px;
+  /* max-width: 848px; */
   /* Figma 본문 영역 너비 제한 */
   display: flex;
   flex-direction: column;
