+++ frontend/src/styles/variables.css
@@ -1,12 +1,12 @@
 /* src/styles/variables.css — 디자인 토큰 (MVP 아키텍처 v3 섹션 10) */
 :root {
   /* Colors */
-  --color-primary: #374151;
-  --color-primary-hover: #6B7280;
+  --color-primary: #172554;
+  --color-primary-hover: #1E3A8A;
   --color-secondary: #4B5563;
   --color-bg: #ffffff;
-  --color-surface: #F3F4F6;
-  --color-surface-hover: #334155;
+  --color-surface: #EFF6FF;
+  --color-surface-hover: #DBEAFE;
 
   --color-text-gray-900: #111827;
   --color-text-gray-700: #374151;
